requirements.txt
lightning==2.2.0
tqdm==4.66.2
pandas==2.2.1
onnx==1.15.0
onnxruntime==1.17.1
onnxsim==0.4.36
tensorboard==2.16.2
torchmetrics==1.3.1
more_itertools==10.2.0
x_transformers==1.27.19